requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosmtplib==5.0.0
alembic==1.17.2
anyio==4.12.0
apache-airflow==2.10.4
apache-airflow-providers-common-compat==1.9.0
apache-airflow-providers-common-io==1.6.5
apache-airflow-providers-common-sql==1.29.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.13.3
apache-airflow-providers-http==5.5.0
apache-airflow-providers-imap==3.9.4
apache-airflow-providers-smtp==2.3.2
apache-airflow-providers-sqlite==4.1.3
apispec==6.9.0
argcomplete==3.6.3
asgiref==3.11.0
attrs==25.4.0
babel==2.17.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
clickclick==20.10.2
colorama==0.4.6
colorlog==6.10.1
ConfigUpdater==3.2
connexion==2.14.2
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.3
Deprecated==1.3.1
dill==0.4.0
dnspython==2.8.0
email-validator==2.3.0
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.8.0
fsspec==2025.12.0
google-re2==1.1.20251105
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
limits==5.6.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.5.0
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
multidict==6.7.0
numpy==2.4.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
pathlib_abc==0.5.2
pathspec==0.12.1
pendulum==3.1.0
pluggy==1.6.0
prison==0.2.1
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.0
psycopg2-binary==2.9.11
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.2
rpds-py==0.30.0
setproctitle==1.3.7
six==1.17.0
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.2.0
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
universal_pathlib==0.3.7
urllib3==2.6.2
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==2.0.1
WTForms==3.2.1
yarl==1.22.0
zipp==3.23.0
